Refactored implementation to use Partial Classes
