+++ tests/conftest.py
@@ -1,9 +1,8 @@
 import pytest
 from testcontainers.postgres import PostgresContainer
 
-from psyscale.client import TIMESCALE_IMAGE, PsyscaleConnectParams
-from psyscale.manager import PsyscaleMod
-
+from psyscale import PsyscaleDB
+from psyscale.core import TIMESCALE_IMAGE, PsyscaleConnectParams
 
 # ---- ---- Module Level Test Container ---- ----
 
@@ -24,4 +23,4 @@ def test_url(test_container):
 @pytest.fixture(scope="module")
 def psyscale_db(test_url):
     "PsyscaleDB Instance that can be reused for the entire session"
-    yield PsyscaleMod(PsyscaleConnectParams.from_url(test_url))
+    yield PsyscaleDB(PsyscaleConnectParams.from_url(test_url))
